requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
celery==5.4.0
charset-normalizer==3.4.3
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.7
django-appconf==1.1.0
django-jazzmin==3.0.3
django-select2==8.4.1
fonttools==4.60.2
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.2
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
redis==5.0.8
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.8.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.11.0
yarl==1.22.0
zopfli==0.2.3.post1
